requirements/local.txt
# Local development dependencies go here
-r base.txt
django-debug-toolbar==1.4